requirements.txt
dj-database-url==1.3.0
Django==4.1.7
django-active-link==0.1.8
django-credit-cards==0.4.1
django-extensions==3.2.1
django-filter==23.1
django-seed==0.3.1
Faker==18.3.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
Pillow==9.4.0
psycopg2==2.9.5